just commit

+++ src/com/company/Function.java
@@ -4,5 +4,5 @@ package com.company;
  * Created by vladstarikov on 21.10.15.
  */
 public interface Function {
-    public float calc(double ... x);
+    public double calc(double ... x);
 }
